adding delete post changes and delete comment api

+++ posts/views.py
@@ -5,6 +5,8 @@ from .serializers import PostSerializer, CommentSerializer, PostCreateSerializer
 from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
 from useraccounts.serializers import ProfileSerializer
+from .permissions import IsOwnerOrReadOnly, has_permission_for_item
+from rest_framework.exceptions import PermissionDenied
 
 
 @api_view(['GET'])
@@ -123,13 +125,44 @@ def add_like_to_post(request, post_id):
 @api_view(['DELETE'])
 @permission_classes([IsAuthenticated])
 def delete_post(request, post_id):
+    """delete post"""
     try:
-        print(post_id)
         post = Post.objects.get(id=post_id)
-        print(post)
+        # print(post)
+        if not has_permission_for_item(request, post):
+            raise PermissionDenied()
         post.delete()
         content = {'message': 'post deleted'}
         return Response(content, status=status.HTTP_200_OK)
+    except Post.DoesNotExist:
+        content = {'message': 'Post Does Not Existed'}
+        return Response(content, status=status.HTTP_404_NOT_FOUND)
+    except PermissionDenied:
+        return Response(PermissionDenied.default_detail, PermissionDenied.status_code)
     except:
         content = {'message': 'failed to delete the post'}
         return Response(content, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
+def delete_comment(request, post_id, comment_id):
+    """delete comment"""
+    try:
+        comment = Comment.objects.get(id=comment_id)
+        if not has_permission_for_item(request, comment):
+            raise PermissionDenied()
+        comment.delete()
+        content = {'message': 'comment deleted'}
+        return Response(content, status=status.HTTP_200_OK)
+    except Comment.DoesNotExist:
+        content = {'message': 'Comment Does Not Existed'}
+        return Response(content, status=status.HTTP_404_NOT_FOUND)
+    except PermissionDenied:
+        content = {'message': PermissionDenied.default_detail}
+        return Response(content, status=PermissionDenied.status_code)
+    except:
+        content = {'message': 'failed to delete the comment'}
+        return Response(content, status=status.HTTP_400_BAD_REQUEST)
+
+
